requirements.txt
flask
opencv-python-headless
fer
Pillow
numpy
